+++ changelog.md
@@ -4,7 +4,7 @@ _(Note: This changelog focuses on the major changes between the different
 versions. Therefore, it may not contain all changes. Especially smaller fixes or
 improvements may be omitted.)_
 
-## Version 0.?.? (2023-05-??)
+## Version 0.6.2 (2023-06-25)
 
 * __[new feature]__
 The Debian package search is extended by adding the `!deb13` command to search